+++ crates/storage/store_db/in_memory.rs
@@ -15,7 +15,7 @@ use std::{
     collections::HashMap,
     fmt::Debug,
     path::Path,
-    sync::{Arc, Mutex, MutexGuard, RwLock},
+    sync::{Arc, Mutex, MutexGuard},
 };
 
 // NOTE: we use a different commit threshold than rocksdb since tests
@@ -39,7 +39,7 @@ pub struct StoreInner {
     // Maps transaction hashes to their blocks (height+hash) and index within the blocks.
     transaction_locations: HashMap<H256, Vec<(BlockNumber, BlockHash, Index)>>,
     receipts: HashMap<BlockHash, HashMap<Index, Receipt>>,
-    trie_cache: Arc<RwLock<TrieLayerCache>>,
+    trie_cache: Arc<TrieLayerCache>,
     // Contains account trie nodes
     state_trie_nodes: NodeMap,
     pending_blocks: HashMap<BlockHash, Block>,
@@ -91,31 +91,28 @@ impl StoreEngine for Store {
         let mut store = self.inner()?;
 
         // Store trie updates
+        let mut trie = TrieLayerCache::clone(&store.trie_cache);
+        let parent = update_batch
+            .blocks
+            .first()
+            .ok_or(StoreError::UpdateBatchNoBlocks)?
+            .header
+            .parent_hash;
+
+        let pre_state_root = store
+            .headers
+            .get(&parent)
+            .map(|header| header.state_root)
+            .unwrap_or_default();
+
+        let last_state_root = update_batch
+            .blocks
+            .last()
+            .ok_or(StoreError::UpdateBatchNoBlocks)?
+            .header
+            .state_root;
+
         {
-            let mut trie = store
-                .trie_cache
-                .write()
-                .map_err(|_| StoreError::LockError)?;
-            let parent = update_batch
-                .blocks
-                .first()
-                .ok_or(StoreError::UpdateBatchNoBlocks)?
-                .header
-                .parent_hash;
-
-            let pre_state_root = store
-                .headers
-                .get(&parent)
-                .map(|header| header.state_root)
-                .unwrap_or_default();
-
-            let last_state_root = update_batch
-                .blocks
-                .last()
-                .ok_or(StoreError::UpdateBatchNoBlocks)?
-                .header
-                .state_root;
-
             let mut state_trie = store
                 .state_trie_nodes
                 .lock()
@@ -131,18 +128,20 @@ impl StoreEngine for Store {
                     }
                 }
             }
-            let key_values = update_batch
-                .storage_updates
-                .into_iter()
-                .flat_map(|(account_hash, nodes)| {
-                    nodes
-                        .into_iter()
-                        .map(move |(path, node)| (apply_prefix(Some(account_hash), path), node))
-                })
-                .chain(update_batch.account_updates)
-                .collect();
-            trie.put_batch(pre_state_root, last_state_root, key_values);
         }
+
+        let key_values = update_batch
+            .storage_updates
+            .into_iter()
+            .flat_map(|(account_hash, nodes)| {
+                nodes
+                    .into_iter()
+                    .map(move |(path, node)| (apply_prefix(Some(account_hash), path), node))
+            })
+            .chain(update_batch.account_updates)
+            .collect();
+        trie.put_batch(pre_state_root, last_state_root, key_values);
+        store.trie_cache = Arc::new(trie);
 
         for block in update_batch.blocks {
             // store block