+++ crates/storage/store_db/rocksdb.rs
@@ -21,7 +21,10 @@ use rocksdb::{
 use std::{
     collections::HashSet,
     path::Path,
-    sync::{Arc, RwLock},
+    sync::{
+        Arc, Mutex,
+        mpsc::{SyncSender, sync_channel},
+    },
 };
 use tracing::{debug, error, info};
 
@@ -116,6 +119,16 @@ pub const CF_FLATKEYVALUE: &str = "flatkeyvalue";
 
 pub const CF_MISC_VALUES: &str = "misc_values";
 
+pub type StorageUpdates = Vec<(H256, Vec<(Nibbles, Vec<u8>)>)>;
+
+pub type TriedUpdateWorkerTx = std::sync::mpsc::SyncSender<(
+    std::sync::mpsc::SyncSender<Result<(), StoreError>>,
+    H256,
+    H256,
+    Vec<(Nibbles, Vec<u8>)>,
+    Vec<(H256, Vec<(Nibbles, Vec<u8>)>)>,
+)>;
+
 /// Control messages for the FlatKeyValue generator
 #[derive(Debug, PartialEq)]
 enum FKVGeneratorControlMessage {
@@ -126,8 +139,9 @@ enum FKVGeneratorControlMessage {
 #[derive(Debug, Clone)]
 pub struct Store {
     db: Arc<DBWithThreadMode<MultiThreaded>>,
-    trie_cache: Arc<RwLock<TrieLayerCache>>,
+    trie_cache: Arc<Mutex<Arc<TrieLayerCache>>>,
     flatkeyvalue_control_tx: std::sync::mpsc::SyncSender<FKVGeneratorControlMessage>,
+    trie_update_worker_tx: TriedUpdateWorkerTx,
 }
 
 impl Store {
@@ -325,16 +339,18 @@ impl Store {
                 }
             }
         }
-        let (tx, rx) = std::sync::mpsc::sync_channel(0);
+        let (fkv_tx, fkv_rx) = std::sync::mpsc::sync_channel(0);
+        let (trie_upd_tx, trie_upd_rx) = std::sync::mpsc::sync_channel(0);
 
         let store = Self {
             db: Arc::new(db),
             trie_cache: Default::default(),
-            flatkeyvalue_control_tx: tx,
+            flatkeyvalue_control_tx: fkv_tx,
+            trie_update_worker_tx: trie_upd_tx,
         };
         let store_clone = store.clone();
         std::thread::spawn(move || {
-            let mut rx = rx;
+            let mut rx = fkv_rx;
             loop {
                 match rx.recv() {
                     Ok(FKVGeneratorControlMessage::Continue) => break,
@@ -351,6 +367,53 @@ impl Store {
                 Err(err) => error!("Error while generating FlatKeyValue: {err}"),
             }
             // rx channel is dropped, closing it
+        });
+        let store_clone = store.clone();
+        /*
+            When a block is executed, the write of the bottom-most diff layer to disk is done in the background through this thread.
+            This is to improve block execution times, since it's not necessary when executing the next block to have this layer flushed to disk.
+
+            This background thread receives messages through a channel to apply new trie updates and does three things:
+
+            - First, it updates the top-most in-memory diff layer and notifies the process that sent the message (i.e. the
+            block production thread) so it can continue with block execution (block execution cannot proceed without the
+            diff layers updated, otherwise it would see wrong state when reading from the trie). This section is done in an RCU manner:
+            a shared pointer with the trie is kept behind a lock. This thread first acquires the lock, then copies the pointer and drops the lock;
+            afterwards it makes a deep copy of the trie layer and mutates it, then takes the lock again, replaces the pointer with the updated copy,
+            then drops the lock again.
+
+            - Second, it performs the logic of persisting the bottom-most diff layer to disk. This is the part of the logic that block execution does not
+            need to proceed. What does need to be aware of this section is the process in charge of generating the snapshot (a.k.a. FlatKeyValue).
+            Because of this, this section first sends a message to pause the FlatKeyValue generation, then persists the diff layer to disk, then notifies
+            again for FlatKeyValue generation to continue.
+
+            - Third, it removes the (no longer needed) bottom-most diff layer from the trie layers in the same way as the first step.
+        */
+        std::thread::spawn(move || {
+            let rx = trie_upd_rx;
+            loop {
+                match rx.recv() {
+                    Ok((
+                        notify,
+                        parent_state_root,
+                        child_state_root,
+                        account_updates,
+                        storage_updates,
+                    )) => {
+                        // FIXME: what should we do on error?
+                        let _ = store_clone
+                            .apply_trie_updates(
+                                notify,
+                                parent_state_root,
+                                child_state_root,
+                                account_updates,
+                                storage_updates,
+                            )
+                            .inspect_err(|err| error!("apply_trie_updates failed: {err}"));
+                    }
+                    Err(err) => error!("Error while reading diff layer: {err}"),
+                }
+            }
         });
         Ok(store)
     }
@@ -607,13 +670,90 @@ impl Store {
             };
         }
     }
+
+    fn apply_trie_updates(
+        &self,
+        notify: SyncSender<Result<(), StoreError>>,
+        parent_state_root: H256,
+        child_state_root: H256,
+        account_updates: Vec<(Nibbles, Vec<u8>)>,
+        storage_updates: StorageUpdates,
+    ) -> Result<(), StoreError> {
+        let db = &*self.db;
+        let fkv_ctl = &self.flatkeyvalue_control_tx;
+        let trie_cache = &self.trie_cache;
+
+        // Phase 1: update the in-memory diff-layers only, then notify block production.
+        let new_layer = storage_updates
+            .into_iter()
+            .flat_map(|(account_hash, nodes)| {
+                nodes
+                    .into_iter()
+                    .map(move |(path, node)| (apply_prefix(Some(account_hash), path), node))
+            })
+            .chain(account_updates)
+            .collect();
+        // Read-Copy-Update the trie cache with a new layer.
+        let trie = trie_cache
+            .lock()
+            .map_err(|_| StoreError::LockError)?
+            .clone();
+        let mut trie_mut = (*trie).clone();
+        trie_mut.put_batch(parent_state_root, child_state_root, new_layer);
+        let trie = Arc::new(trie_mut);
+        *trie_cache.lock().map_err(|_| StoreError::LockError)? = trie.clone();
+        // Update finished, signal block processing.
+        notify.send(Ok(())).map_err(|_| StoreError::LockError)?;
+
+        // Phase 2: update disk layer.
+        let Some(root) = trie.get_commitable(parent_state_root, COMMIT_THRESHOLD) else {
+            // Nothing to commit to disk, move on.
+            return Ok(());
+        };
+        // Stop the flat-key-value generator thread, as the underlying trie is about to change.
+        // Ignore the error, if the channel is closed it means there is no worker to notify.
+        let _ = fkv_ctl.send(FKVGeneratorControlMessage::Stop);
+
+        // RCU to remove the bottom layer: update step needs to happen after disk layer is updated.
+        let mut trie_mut = (*trie).clone();
+        let mut batch = WriteBatch::default();
+        let [cf_trie_nodes, cf_flatkeyvalue, cf_misc] =
+            open_cfs(db, [CF_TRIE_NODES, CF_FLATKEYVALUE, CF_MISC_VALUES])?;
+
+        let last_written = db.get_cf(&cf_misc, "last_written")?.unwrap_or_default();
+        // Commit removes the bottom layer and returns it, this is the mutation step.
+        let nodes = trie_mut.commit(root).unwrap_or_default();
+        for (key, value) in nodes {
+            let is_leaf = key.len() == 65 || key.len() == 131;
+
+            if is_leaf && key > last_written {
+                continue;
+            }
+            let cf = if is_leaf {
+                &cf_flatkeyvalue
+            } else {
+                &cf_trie_nodes
+            };
+            if value.is_empty() {
+                batch.delete_cf(cf, key);
+            } else {
+                batch.put_cf(cf, key, value);
+            }
+        }
+        let result = db.write(batch);
+        // We want to send this message even if there was an error during the batch write
+        let _ = fkv_ctl.send(FKVGeneratorControlMessage::Continue);
+        result?;
+        // Phase 3: update diff layers with the removal of bottom layer.
+        *trie_cache.lock().map_err(|_| StoreError::LockError)? = Arc::new(trie_mut);
+        Ok(())
+    }
 }
 
 #[async_trait::async_trait]
 impl StoreEngine for Store {
     async fn apply_updates(&self, update_batch: UpdateBatch) -> Result<(), StoreError> {
         let db = self.db.clone();
-        let trie_cache = self.trie_cache.clone();
         let parent_state_root = self
             .get_block_header_by_hash(
                 update_batch
@@ -631,80 +771,52 @@ impl StoreEngine for Store {
             .ok_or(StoreError::UpdateBatchNoBlocks)?
             .header
             .state_root;
-        let flatkeyvalue_control_tx = self.flatkeyvalue_control_tx.clone();
+        let trie_upd_worker_tx = self.trie_update_worker_tx.clone();
 
         tokio::task::spawn_blocking(move || {
             let _span = tracing::trace_span!("Block DB update").entered();
 
             let [
-                cf_trie_nodes,
-                cf_flatkeyvalue,
                 cf_receipts,
                 cf_codes,
                 cf_block_numbers,
                 cf_tx_locations,
                 cf_headers,
                 cf_bodies,
-                cf_misc,
             ] = open_cfs(
                 &db,
                 [
-                    CF_TRIE_NODES,
-                    CF_FLATKEYVALUE,
                     CF_RECEIPTS,
                     CF_ACCOUNT_CODES,
                     CF_BLOCK_NUMBERS,
                     CF_TRANSACTION_LOCATIONS,
                     CF_HEADERS,
                     CF_BODIES,
-                    CF_MISC_VALUES,
                 ],
             )?;
-
+            let _span = tracing::trace_span!("Block DB update").entered();
             let mut batch = WriteBatch::default();
 
-            let mut updated_trie = false;
+            let UpdateBatch {
+                account_updates,
+                storage_updates,
+                ..
+            } = update_batch;
 
-            let mut trie = trie_cache.write().map_err(|_| StoreError::LockError)?;
-            if let Some(root) = trie.get_commitable(parent_state_root, COMMIT_THRESHOLD) {
-                updated_trie = true;
-                // If the channel is closed, there's nobody to notify
-                let _ = flatkeyvalue_control_tx.send(FKVGeneratorControlMessage::Stop);
-
-                let last_written = db.get_cf(&cf_misc, "last_written")?.unwrap_or_default();
-                let nodes = trie.commit(root).unwrap_or_default();
-                for (key, value) in nodes {
-                    let is_leaf = key.len() == 65 || key.len() == 131;
-
-                    if is_leaf && key > last_written {
-                        continue;
-                    }
-                    let cf = if is_leaf {
-                        &cf_flatkeyvalue
-                    } else {
-                        &cf_trie_nodes
-                    };
-                    if value.is_empty() {
-                        batch.delete_cf(cf, key);
-                    } else {
-                        batch.put_cf(cf, key, value);
-                    }
-                }
-            }
-            trie.put_batch(
-                parent_state_root,
-                last_state_root,
-                update_batch
-                    .storage_updates
-                    .into_iter()
-                    .flat_map(|(account_hash, nodes)| {
-                        nodes
-                            .into_iter()
-                            .map(move |(path, node)| (apply_prefix(Some(account_hash), path), node))
-                    })
-                    .chain(update_batch.account_updates)
-                    .collect(),
-            );
+            // Capacity one ensures sender just notifies and goes on
+            let (notify_tx, notify_rx) = sync_channel(1);
+            let wait_for_new_layer = notify_rx;
+            trie_upd_worker_tx
+                .send((
+                    notify_tx,
+                    parent_state_root,
+                    last_state_root,
+                    account_updates,
+                    storage_updates,
+                ))
+                .map_err(|e| {
+                    StoreError::Custom(format!("failed to read new trie layer notification: {e}"))
+                })?;
 
             for block in update_batch.blocks {
                 let block_number = block.header.number;
@@ -753,15 +865,14 @@ impl StoreEngine for Store {
                 batch.put_cf(&cf_codes, code_hash.0, buf);
             }
 
-            // Single write operation
-            let ret = db
-                .write(batch)
-                .map_err(|e| StoreError::Custom(format!("RocksDB batch write error: {}", e)));
-            if updated_trie {
-                // If the channel is closed, there's nobody to notify
-                let _ = flatkeyvalue_control_tx.send(FKVGeneratorControlMessage::Continue);
-            }
-            ret
+            // Wait for an updated top layer so every caller afterwards sees a consistent view.
+            // Specifically, the next block produced MUST see this upper layer.
+            wait_for_new_layer
+                .recv()
+                .map_err(|e| StoreError::Custom(format!("recv failed: {e}")))??;
+            // After top-level is added, we can make the rest of the changes visible.
+            db.write(batch)
+                .map_err(|e| StoreError::Custom(format!("RocksDB batch write error: {}", e)))
         })
         .await
         .map_err(|e| StoreError::Custom(format!("Task panicked: {}", e)))?
@@ -1329,10 +1440,15 @@ impl StoreEngine for Store {
         storage_root: H256,
         state_root: H256,
     ) -> Result<Trie, StoreError> {
+        // FIXME: use a DB snapshot here
         let db = Box::new(RocksDBTrieDB::new(self.db.clone(), CF_TRIE_NODES, None)?);
         let wrap_db = Box::new(TrieWrapper {
             state_root,
-            inner: self.trie_cache.clone(),
+            inner: self
+                .trie_cache
+                .lock()
+                .map_err(|_| StoreError::LockError)?
+                .clone(),
             db,
             prefix: Some(hashed_address),
         });
@@ -1340,10 +1456,15 @@ impl StoreEngine for Store {
     }
 
     fn open_state_trie(&self, state_root: H256) -> Result<Trie, StoreError> {
+        // FIXME: use a DB snapshot here
         let db = Box::new(RocksDBTrieDB::new(self.db.clone(), CF_TRIE_NODES, None)?);
         let wrap_db = Box::new(TrieWrapper {
             state_root,
-            inner: self.trie_cache.clone(),
+            inner: self
+                .trie_cache
+                .lock()
+                .map_err(|_| StoreError::LockError)?
+                .clone(),
             db,
             prefix: None,
         });
@@ -1376,7 +1497,11 @@ impl StoreEngine for Store {
         )?);
         let wrap_db = Box::new(TrieWrapper {
             state_root,
-            inner: self.trie_cache.clone(),
+            inner: self
+                .trie_cache
+                .lock()
+                .map_err(|_| StoreError::LockError)?
+                .clone(),
             db,
             prefix: None,
         });
@@ -1396,7 +1521,11 @@ impl StoreEngine for Store {
         )?);
         let wrap_db = Box::new(TrieWrapper {
             state_root,
-            inner: self.trie_cache.clone(),
+            inner: self
+                .trie_cache
+                .lock()
+                .map_err(|_| StoreError::LockError)?
+                .clone(),
             db,
             prefix: Some(hashed_address),
         });
@@ -1780,7 +1909,7 @@ impl StoreEngine for Store {
 
 /// Open column families
 fn open_cfs<'a, const N: usize>(
-    db: &'a Arc<DBWithThreadMode<MultiThreaded>>,
+    db: &'a DBWithThreadMode<MultiThreaded>,
     names: [&str; N],
 ) -> Result<[Arc<BoundColumnFamily<'a>>; N], StoreError> {
     let mut handles = Vec::with_capacity(N);